refactor: ReservationStatusService 기능 ReservationService로 통합
