+++ src/main/java/roomescape/reservation/application/ReservationService.java
@@ -3,6 +3,7 @@ package roomescape.reservation.application;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
+import java.util.Arrays;
 import java.util.List;
 import java.util.Objects;
 import lombok.RequiredArgsConstructor;
@@ -24,6 +25,7 @@ import roomescape.reservation.ui.dto.request.CreateReservationRequest;
 import roomescape.reservation.ui.dto.request.ReservationsByFilterRequest;
 import roomescape.reservation.ui.dto.response.AvailableReservationTimeResponse;
 import roomescape.reservation.ui.dto.response.ReservationResponse;
+import roomescape.reservation.ui.dto.response.ReservationStatusResponse;
 import roomescape.theme.domain.Theme;
 import roomescape.theme.domain.ThemeQueryRepository;
 
@@ -162,6 +164,12 @@ public class ReservationService {
     public List<ReservationResponse.ForMember> findReservationsByMemberId(final Long memberId) {
         return reservationQueryRepository.findAllByMemberId(memberId).stream()
                 .map(ReservationResponse.ForMember::from)
+                .toList();
+    }
+
+    public List<ReservationStatusResponse> findAllReservationStatuses() {
+        return Arrays.stream(ReservationStatus.values())
+                .map(ReservationStatusResponse::from)
                 .toList();
     }
 }
