+++ src/main/java/roomescape/reservation/ui/ReservationRestController.java
@@ -22,6 +22,7 @@ import roomescape.reservation.ui.dto.request.AvailableReservationTimeRequest;
 import roomescape.reservation.ui.dto.request.CreateReservationRequest;
 import roomescape.reservation.ui.dto.response.AvailableReservationTimeResponse;
 import roomescape.reservation.ui.dto.response.ReservationResponse;
+import roomescape.reservation.ui.dto.response.ReservationStatusResponse;
 
 @RestController
 @RequiredArgsConstructor
@@ -69,5 +70,12 @@ public class ReservationRestController {
                 reservationService.findAvailableReservationTimes(request);
 
         return ResponseEntity.ok(availableReservationTimes);
+    }
+
+    @GetMapping("/reservation-statuses")
+    @RequiresRole(authRoles = {ADMIN, MEMBER})
+    public ResponseEntity<List<ReservationStatusResponse>> findAll() {
+        return ResponseEntity.ok()
+                .body(reservationService.findAllReservationStatuses());
     }
 }
