+++ src/test/java/roomescape/fixture/config/TestConfig.java
@@ -9,7 +9,6 @@ import roomescape.member.domain.MemberQueryRepository;
 import roomescape.member.infrastructure.JpaMemberRepository;
 import roomescape.member.infrastructure.MemberRepositoryImpl;
 import roomescape.reservation.application.ReservationService;
-import roomescape.reservation.application.ReservationStatusService;
 import roomescape.reservation.application.ReservationTimeService;
 import roomescape.reservation.domain.ReservationCommandRepository;
 import roomescape.reservation.domain.ReservationQueryRepository;
@@ -123,10 +122,5 @@ public class TestConfig {
             final MemberQueryRepository memberQueryRepository
     ) {
         return new MemberService(memberCommandRepository, memberQueryRepository);
-    }
-
-    @Bean
-    public ReservationStatusService reservationStatusService() {
-        return new ReservationStatusService();
     }
 }
